backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas",email = "afonsos.lucassoares@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "django-ninja (>=1.5.3,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "scalar-ninja (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
